requirements.txt
python-coinpayments@git+https://github.com/HK-Mattew/python-coinpayments.git